some more cosmetical changes
