+++ src/main/groovy/org/aim42/htmlsanitycheck/ChecksRunner.groovy
@@ -1,6 +1,6 @@
 package org.aim42.htmlsanitycheck
 
-import org.aim42.htmlsanitycheck.check.BrokenCrossReferencesChecker
+import org.aim42.filesystem.FileUtil
 import org.aim42.htmlsanitycheck.check.Checker
 import org.aim42.htmlsanitycheck.check.CheckerCreator
 import org.aim42.htmlsanitycheck.collect.PerRunResults
@@ -13,8 +13,6 @@ import org.aim42.htmlsanitycheck.report.LoggerReporter
 import org.aim42.htmlsanitycheck.report.Reporter
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-
-import java.nio.file.Path
 
 /**
  * runs one or several checks on HTML input
@@ -104,39 +102,12 @@ class ChecksRunner {
         this.checkingResultsDir = checkingResultsDir
 		this.junitResultsDir = junitResultsDir
 
-		def params = [baseDirPath: commonPath(filesToCheck).canonicalPath]
+		def params = [baseDirPath: FileUtil.commonPath(filesToCheck).canonicalPath]
 
         this.checkers = CheckerCreator.createCheckerClassesFrom( checkerCollection, params )
 
         logger.debug("ChecksRunner created with ${checkerCollection.size()} checkers for ${filesToCheck.size()} files")
     }
-
-	static File commonPath(Collection<File> files) {
-		if (!files) {
-			return null
-		}
-		if (files.size() == 1) {
-			return files.first().parentFile
-		}
-        Path initial = files.first().toPath().parent
-		Path common = files.collect { it.toPath() }.inject(initial) { Path acc, Path val ->
-			if (!acc || !val) {
-				return null
-			}
-			
-			int idx = 0
-			Path p1 = acc, p2 = val.parent
-			def iter1 = p1.iterator(), iter2 = p2.iterator()
-			while (iter1.hasNext() && iter2.hasNext() && iter1.next() == iter2.next()) {
-				idx++
-			}
-			if (idx == 0) {
-				return null
-			}
-			p1.subpath(0, idx)
-		}
-        common ? (initial?.root ? initial.root.resolve(common).toFile() : common.toFile()) : null
-	}
 
     /**
      *  performs all configured checks on a single HTML file.
